requirements.txt
paho-mqtt==1.6.1
python-roborock==2.6.0
